+++ Projects/Modules/Networks/Sources/DTO/My/MyPageDTO.swift
@@ -8,36 +8,16 @@
 
 import Foundation
 
+public struct MyProfileResult: Decodable {
+    public let nickname: String
+    public let university: String
+    public let profileImageUrl: String
+}
+
 public struct EditNameRequest: Encodable {
     let nickName: String
     
     public init(_ nickName: String) {
         self.nickName = nickName
     }
-}
-
-public struct AcademicDTO: Codable {
-    let matriculationYear: Int
-    let academicYear: Int
-    let studentId: String
-    
-    public init(
-        _ matriculationYear: Int,
-        _ academicYear: Int,
-        _ studentId: String
-    ) {
-        self.matriculationYear = matriculationYear
-        self.academicYear = academicYear
-        self.studentId = studentId
-    }
-}
-
-public struct profileImgResult: Decodable {
-    let imageUrl: String
-}
-
-public struct MyProfileResult: Decodable {
-    public let nickname: String
-    public let university: String
-    public let profileImageUrl: String
 }
